Adding typescript and type Filter exposed from App1
