+++ app1/package.json
@@ -8,6 +8,10 @@
     "build": "webpack --mode production"
   },
   "dependencies": {
+    "@types/jest": "^27.4.0",
+    "@types/node": "^17.0.7",
+    "@types/react": "^17.0.38",
+    "@types/react-dom": "^17.0.11",
     "react": "^17.0.2",
     "react-dom": "^17.0.2"
   },
@@ -17,6 +21,8 @@
     "babel-loader": "^8.2.2",
     "external-remotes-plugin": "^1.0.0",
     "html-webpack-plugin": "^5.3.1",
+    "ts-loader": "^9.2.6",
+    "typescript": "^4.5.4",
     "webpack": "^5.38.1",
     "webpack-cli": "^4.7.2",
     "webpack-dev-server": "^4.0.0"
