+++ app1/src/Filter.tsx
@@ -1,13 +1,17 @@
 import * as React from 'react';
 
-const Filter = () => {
+interface FilterProps {
+    project: string;
+}
+
+const Filter: React.FC<FilterProps> = ({project}) => {
     return (<div style={{
         margin:"10px",
         padding:"10px",
         textAlign:"center",
         backgroundColor:"blueviolet"
     }}>
-        <h1>Filters</h1>
+        <h1>Filters - {project}</h1>
     </div>)
 }
