+++ app1/src/bootstrap.tsx
@@ -1,5 +1,5 @@
-import App from "./App";
 import React from "react";
 import ReactDOM from "react-dom";
+import App from "./App";
 
 ReactDOM.render(<App/>, document.getElementById("root"));
