+++ app1/src/modules/app2.d.ts
@@ -1,0 +1,1 @@
+declare module 'app2/App';
