+++ app1/src/modules/app3.d.ts
@@ -1,0 +1,1 @@
+declare module 'app3/App';
