+++ app2/src/index.ts
@@ -1,3 +1,4 @@
+// @ts-ignore
 window.app1Url = "http://localhost:3001"
 
 import("./bootstrap");
