+++ app3/src/modules/app1.d.ts
@@ -1,0 +1,1 @@
+declare module 'app1/Filter';
